Refactor AllDataComponent into it's own file
